+++ spring-projects/trainee-app-with-mongodb/src/main/java/com/ust/traineeapp/service/TraineeService.java
@@ -2,19 +2,20 @@ package com.ust.traineeapp.service;
 
 import com.ust.traineeapp.model.Trainee;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface TraineeService {
 
     Trainee saveTrainee(Trainee trainee);
 
-    Trainee findTraineeById(int id);
+    Trainee findTraineeById(BigInteger id);
 
-    void removeTrainee(int id);
+    void removeTrainee(BigInteger id);
 
     List<Trainee> getAllTrainees();
 
-    Trainee updateTrainee(int id, Trainee trainee);
+    Trainee updateTrainee(BigInteger id, Trainee trainee);
 
 
 }
